requirements.txt
httpx>=0.27
asyncpg>=0.30
fastembed>=0.4
playwright>=1.49
rich>=13.9
pyyaml>=6.0
duckduckgo-search>=7.0
fastapi>=0.115
uvicorn[standard]>=0.32
websockets>=14.0
neo4j>=5.0
